requirements.txt
groq
python-dotenv
pdfplumber
PyPDF2
requests
pymongo
numpy
fastapi
uvicorn
python-multipart